refactor(coderd/templatebuilder): extract validation helpers to validate.go
Move validation, HCL quoting, and type-checking functions from compose.go
into a dedicated validate.go file to keep compose.go focused on the
compose pipeline. Corresponding tests move to validate_internal_test.go.

No functional changes.
